ai-engine/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
setuptools>=68
wheel>=0.41
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.26.4
torch==2.5.1+cpu
transformers==4.54.1
joblib==1.3.2
safetensors==0.4.5
google-generativeai==0.3.2
scikit-learn==1.4.2
